backend/pyproject.toml
[project]
name = "compute-oracle"
version = "0.1.0"
description = "Self-improving agent that predicts compute cost fluctuations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis[hiredis]>=5.2.0",
    "boto3>=1.35.0",
    "openai>=1.55.0",
    "weave>=0.51.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "stagehand>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
